refactor: changed record.py to record_utils.py
