+++ src/caret_analyze/record/__init__.py
@@ -14,12 +14,11 @@
 
 from .column import Column, Columns, ColumnValue
 from .data_frame_shaper import Clip, DataFrameShaper, Strip
-from .record import (merge,
-                     merge_sequential,
-                     merge_sequential_for_addr_track,
-                     RecordInterface,
-                     RecordsInterface)
+from .record_cpp_impl import RecordInterface, RecordsInterface
 from .record_factory import RecordFactory, RecordsFactory
+from .record_utils import (merge,
+                           merge_sequential,
+                           merge_sequential_for_addr_track)
 from .records_service import Frequency, Latency, Period, Range, ResponseTime, StackedBar
 
 __all__ = [
