+++ src/caret_analyze/record/record_utils.py
@@ -17,7 +17,7 @@ from __future__ import annotations
 from enum import IntEnum
 from itertools import groupby
 
-from .interface import RecordsInterface, RecordInterface
+from .interface import RecordsInterface
 from ..exceptions import InvalidArgumentError
