+++ src/caret_analyze/record/records_service/stacked_bar.py
@@ -17,8 +17,8 @@ from __future__ import annotations
 from collections.abc import Sequence
 
 from .latency import Latency
-from ..interface import RecordsInterface
 from ..column import ColumnValue
+from ..interface import RecordsInterface
 from ..record_factory import RecordsFactory
 from ...common import ClockConverter
